Create SQL_HC_Template.py
Use this script as a template to display whatever the SQL queries want to display for the alerting place the query in the function and mentioned column list in the html output object.  If you are displaying 5 columns, then output should be 5 column list in the output object. In this script shown few.